See comment
Removed Liveries until we come up with a solution to not have triplicates
Added Luas to enable tuning of GBU 10, 12 and 16.
Added new textures for Meteor and Stormshadow
Fixed Shape model for 1700 L tank
Removed files that were not in use. Bomb type 200 and some input profiles
Removed content of jsow_bk_90.lua in preparation of cbu 97 reskin